+++ CHANGES.rst
@@ -12,6 +12,10 @@ Changelog
 - Partly remove Python2 compatibility.
   [cekk]
   
+- Fix SolrIndexProcessor logic: avoid not needed reindexes when we are reindexing objects with indexes not in SOLR.
+  [cekk]
+
+
 1.4.2 (2024-08-08)
 ------------------